Pulled SpecialFormDeclarations out of DebugDeclarations.
